Refactor our Anchor component
