+++ src/components/PropertySubsection/caseDetailSubsection.js
@@ -1,27 +1,7 @@
 // Component to display a property
 import { Grid, withStyles } from '@material-ui/core';
-import { Link } from 'react-router-dom';
 import React from 'react';
-
-function prepareLinks(properties, data) {
-  return properties.map((prop) => {
-    const newProp = { ...prop };
-    const pattern = /{(.*)}/;
-    if (prop.link) {
-      newProp.link = prop.link.replace(pattern, (match, p1) => data[p1]);
-    }
-    if (prop.labelLink) {
-      newProp.labelLink = prop.labelLink.replace(pattern, (match, p1) => data[p1]);
-    }
-    return newProp;
-  });
-}
-
-const Anchor = ({ link, text, classes }) => (
-  link.match(/\w+:\/\//)
-    ? <a href={link} target="_blank" rel="noopener noreferrer" className={classes.link}>{text}</a>
-    : <Link to={link} className={classes.link}>{text}</Link>
-);
+import { Anchor, prepareLinks } from '../Anchor/anchor';
 
 const PropertyItem = ({
   label, value, link, labelLink, classes,
